+++ fixture-app/tests/AllRoutesTest.php
@@ -3,9 +3,7 @@
 namespace App\Tests;
 
 use Pierstoval\SmokeTesting\SmokeTestStaticRoutes;
-use Symfony\Bundle\FrameworkBundle\Test\WebTestCase;
 
-class AllRoutesTest extends WebTestCase
+class AllRoutesTest extends SmokeTestStaticRoutes
 {
-    use SmokeTestStaticRoutes;
 }